add standartized responsees for each endpoint and check some logical case in business logic
-----------------

modified:   TestTask.API/Controllers/MoneyTransferController.cs modified:   TestTask.API/Controllers/UserController.cs modified:   TestTask.API/Controllers/WalletController.cs modified:   TestTask.Application/Commands/MoneyTransfers/TransferMoneyCommandHandler.cs deleted:    TestTask.Application/Commands/MoneyTransfers/TransferMonneyCommand.cs modified:   TestTask.Application/Commands/Users/CreateUserCommand.cs modified:   TestTask.Application/Commands/Users/CreateUserCommandHandler.cs modified:   TestTask.Application/Commands/Users/DeleteUserCommand.cs modified:   TestTask.Application/Commands/Users/DeleteUserCommandHandler.cs modified:   TestTask.Application/Commands/Wallet/CreateWalletCommand.cs modified:   TestTask.Application/Commands/Wallet/CreateWalletCommandHandler.cs modified:   TestTask.Domain/Interfaces/IUserRepository.cs modified:   TestTask.Infrastructure/Persistence/Repositories/UserRepository.cs

+++ TestTask.Application/Commands/Users/DeleteUserCommandHandler.cs
@@ -4,11 +4,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using TestTask.Application.DTOs.Responses;
 using TestTask.Domain.Interfaces;
 
 namespace TestTask.Application.Commands.Users
 {
-    public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, Unit>
+    public class DeleteUserCommandHandler : IRequestHandler<DeleteUserCommand, BaseResponse<DeleteUserResponse>>
     {
         private readonly IUnitOfWork _unitOfWork;
 
@@ -17,11 +18,32 @@ namespace TestTask.Application.Commands.Users
             _unitOfWork = unitOfWork;
         }
 
-        public async Task<Unit> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+        public async Task<BaseResponse<DeleteUserResponse>> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
         {
-            await _unitOfWork.UserRepository.RemoveAsync(request.UserId);
-            await _unitOfWork.SaveChangesAsync();
-            return Unit.Value;
+            try
+            {
+                var user = await _unitOfWork.UserRepository.GetByIdAsync(request.UserId);
+
+                if (user == null)
+                {
+                    return new BaseResponse<DeleteUserResponse>(false, "User not found.", null);
+                }
+
+                await _unitOfWork.UserRepository.RemoveAsync(request.UserId);
+                await _unitOfWork.SaveChangesAsync();
+
+                var response = new DeleteUserResponse
+                {
+                    UserId = request.UserId
+                };
+
+                return new BaseResponse<DeleteUserResponse>(true, "User deleted successfully.", response);
+            }
+            catch (Exception ex)
+            {
+                // Log the exception
+                return new BaseResponse<DeleteUserResponse>(false, ex.Message, null);
+            }
         }
     }
 }
